requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
dnspython==2.6.1
fastapi==0.112.2
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
idna==3.8
joblib==1.4.2
motor==3.5.1
numpy==2.1.0
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.3
strawberry-graphql==0.239.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
